+++ index.js
@@ -3,7 +3,7 @@
 // Makes the script crash on unhandled rejections instead of silently
 // ignoring them. In the future, promise rejections that are not handled will
 // terminate the Node.js process with a non-zero exit code.
-process.on('unhandledRejection', (err) => {
+process.on('unhandledRejection', err => {
 	handleError(`UNHANDLED ERROR`, err);
 });
 
@@ -16,6 +16,7 @@ const theEnd = require('./utils/theEnd.js');
 const handleError = require('cli-handle-error');
 const getStates = require('./utils/getStates.js');
 const getCountry = require('./utils/getCountry.js');
+const getCountryChart = require('./utils/getCountryChart.js');
 const getWorldwide = require('./utils/getWorldwide.js');
 const getCountries = require('./utils/getCountries.js');
 const getCountriesCompare = require('./utils/getCountriesCompare.js');
@@ -27,18 +28,22 @@ const {
 	coloredStates,
 	borderless
 } = require('./utils/table.js');
+
+// Cli.
+const [input] = cli.input;
 const xcolor = cli.flags.xcolor;
 const sortBy = cli.flags.sort;
 const reverse = cli.flags.reverse;
 const limit = Math.abs(cli.flags.limit);
+const chart = cli.flags.chart;
+const log = cli.flags.log;
 const minimal = cli.flags.minimal;
 const compare = cli.flags.compare.length > 1;
-const options = { sortBy, limit, reverse, minimal, compare };
+const options = { sortBy, limit, reverse, minimal, chart, log, compare };
 
 (async () => {
 	// Init.
 	init(minimal);
-	const [input] = cli.input;
 	input === 'help' && (await cli.showHelp(0));
 	const states = input === 'states' ? true : false;
 	const country = input;
@@ -60,6 +65,7 @@ const options = { sortBy, limit, reverse, minimal, compare };
 	await getCountriesCompare(spinner, table, compareCountries, options);
 	await getStates(spinner, table, states, options);
 	await getCountries(spinner, table, states, country, options);
+	await getCountryChart(spinner, country, options);
 
 	theEnd(lastUpdated, states, minimal);
 })();