+++ utils/getCountries.js
@@ -22,9 +22,6 @@ module.exports = async (
 		handleError(`API is down, try again later.`, err, false);
 		let allCountries = response.data;
 
-		// Limit.
-		allCountries = allCountries.slice(0, limit);
-
 		// Sort & reverse.
 		const direction = reverse ? 'asc' : 'desc';
 		allCountries = orderBy(
@@ -32,6 +29,9 @@ module.exports = async (
 			[sortingKeys[sortBy]],
 			[direction]
 		);
+
+		// Limit.
+		allCountries = allCountries.slice(0, limit);
 
 		// Push selected data.
 		allCountries.map((oneCountry, count) => {